Finally, the merge down from tt-processController.
This merge reflects the efforts we made on the TT project, bringing the
fork down back into what is essentially them main branch - rh6pd.

The merge is not complicated - no refactoring in other projects really
necessary.

FIX: Made an improvement to Environment.hasDefinition, to check the
definition ID. Unit tests should assert on this method for a clean fail
if the environment is not setup properly.

Kudos to ajarret for preparing a good repo expoert in
tt-processController/src/test/resources (included in merge).